bug(create-entry-middleware-file): Create class EntryMiddleware to contain all middlewares relating to the entry controller - moves all entry middleware methods to EntryMiddleware class - instantiate EntryMiddleware class in route file - call the various methods of the EntryMiddleware class on the newly created object in the neccessary API endpoints in route file - delete the auth file that contains all middleware methods [Fixes #159502828]
