+++ test/users.js
@@ -1,7 +1,7 @@
 import pg from 'pg';
 import supertest from 'supertest';
 import { expect } from 'chai';
-import app from '../index';
+import app from '..';
 
 const request = supertest(app);
 const connectionString = process.env.DATABASE_URL || 'postgres://postgres:postgres@localhost:5432/mydiary_dev';
